Restructuring
Commented by John and Graeme make the arguments more concise.

Reorganise the material in Sec 4 to present things in a more
coherent way, organised by topic. Reduce duplication of points
in places.

Remove "Roadmap" as a section 4 title as this is not
really what's prestented. Point to continuing work in the
auspices of the HSF Training WG in the conclusion.